refactorlayout naming
