+++ app/src/layouts/RecruiterLayout.vue
